test: move component tests under unit scope
