+++ tests/unit/artifacts/test_exceptions.py
