+++ tests/unit/artifacts/test_manifest_emission.py
